Several minor improvements to the app
Compute the window-in-collection state instead of storing in store
Only save current window to collection and don't overwrite other unsaved
windows
Stop toggling variable dialog when saving to collection
